project-1-shell/requirements.txt
matplotlib==3.8.1
python_igraph==0.11.2
